Consolidate duplicated utilities and optimize SQL parse functions
- Extract normalizeCodeType (3 copies) and normalizeCodeValue (2 copies) into
  shared body-site-laterality-constants module
- Deduplicate extractLaterality/extractBodySite with generic extractFromSet<T>
- Extract shared backfill runner (scripts/lib/run-backfill.ts) from ~240 lines
  duplicated between backfill-laterality.ts and backfill-body-site.ts
- Use CTE-based SQL in backfill scripts to call parse function once per row
  instead of twice (WHERE + SET)
- Rewrite parse_laterality and parse_body_site as plpgsql to hoist upper()
  into a local variable (20x → 1x per row)
- Bundle laterality/bodySite into LookupFilters object in lookup.ts to reduce
  parameter sprawl through the query chain
